restored u16 to pc

+++ src/intel8080_state.rs
@@ -12,6 +12,6 @@ pub struct StateIntel8080
     pub h: u8,
     pub l: u8,
     pub sp: u16,
-    pub pc: u8,
+    pub pc: u16,
     pub condition: ConditionCodes
 }
